fix(ci): file line limit — split oversized executor coverage part_009 into part_020
